requirements.txt
absl-py==1.4.0
brotlipy==0.7.0
cachetools==5.3.0
cycler==0.11.0
fonttools==4.38.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.3
importlib-metadata==6.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-service==2.4.0
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
sklearn==0.0.post1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==3.1.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.1
Werkzeug==2.2.3
zipp==3.15.0
